Have 2 kinds of reports: at constructor, and in-situ.
Since each kind of near miss is solved in different places (at the
constructor, and at or "near" the failure site, respectively), makes
sense to consider those to be separate.

Different kinds of merging are done for each kind. For example,
constructor reports merge failures about all fields into a single report
(because they could be fixed in the same place / at the same time), but
in-situ reports emit one report per field (each field is unrelated if
you fix near misses at the failure site).

This leads to less spurious merging, where low-badness fields would end
up in reports for near misses that wouldn't be fixed at the constructor.
This in turn leads to better ranking/pruning, because more in-situ near
misses are considered separate and don't end up clogging up the hot
reports with mega-reports about unrelated issues.

This also allows different information to be shown for each kind of
report, which makes it possible to show all the necessary information to
be actionable, while omitting irrelevant info.